Geometry Script, Scraper
The geometry for the simulation was created in another script and then executed in the main script. This reduces the clutter of the program.

A scraper function was added that will delete particles that reach a certain extent. This allowed for significant modifications in the plotting profile which now doesnt have to create seperate dataframes to plot each particle. The plotting routine now runs much faster and there is less room for error.

Also, a more reliable count of lost particles is now available with the scraper deleting particles from the simulation.