Animation using Blit
Test file was renamed to Blit_animation_examples.py. This scripts contains multiple steps into using Blit for creating animations. The end goal was to create an animated scatter plot on multiple pltos. This was acheived. However, it took multiple iterations to finally acheive this since the technique is no obvious.