Created R column for animations
A column of r-values (sqrt(x^2 + y^2)) was added to the dataframe for post processing.

The animation now uses r instead of x. This is more optimized since it caputres x-only coordinates too.

The initial r animation starts off with particles above the scraper and then immediately scrapes them. This should be investigated. Perhaps the scaper is not being called soon enoughor the trajectories can be written after the initial scrape.

Realistic gaussian distribution for 4000 particles was simulated and animated. Simulation time for 4-bounces took 59 seconds and simulation took a 1min 15 seconds. Not bad.