Simple Animation, Geometry Addition
The simple statistics annimation is now good to go. It will give sigmas vs z and t along with the ray tracing of the particle. These can easily be changed to give different measures or visualize particle movement rather than ray tracing.

The end wall were added to the geometry file on the right and left. I made an annulus with a 25mm hole and set them each an electrode length way away.